requirements.txt
twilio
google-cloud-vision
gspread
oauth2client
openai
requests
python-dotenv
flask
gunicorn
flask-limiter